# high-load-systems/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CDN Control Plane API",
    "description": "\nControl plane for a content delivery network: the service that decides **where**\ncontent lives and **which** edge node a client should talk to. It never serves the\ncontent bytes itself.\n\n* **Node management** — registration, lifecycle and continuous telemetry ingest.\n* **Routing** — ranked edge-node selection by proximity, load and cache state.\n* **Distribution** — asset metadata and the regional policy governing where it may be cached.\n* **Invalidation** — tracked purge campaigns with per-node acknowledgement.\n\nErrors follow RFC 9457 and are returned as `application/problem+json`.\nEvery response carries `X-Instance-ID`, `X-Request-ID` and `X-Response-Time-ms`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/purges": {
      "get": {
        "tags": [
          "purges"
        ],
        "summary": "List purge campaigns",
        "operationId": "list_purges_api_v1_purges_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PurgeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PurgeRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "purges"
        ],
        "summary": "Broadcast a cache invalidation",
        "description": "Open a purge campaign against the nodes in scope.\n\n202, not 201: the campaign record is created synchronously, but the invalidation\nitself only completes when the targeted nodes acknowledge it. Returning 201 would\nclaim the content is already gone from the edge.",
        "operationId": "create_purge_api_v1_purges_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation or violated a business rule.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purges/{purge_id}": {
      "get": {
        "tags": [
          "purges"
        ],
        "summary": "Get purge progress",
        "operationId": "get_purge_api_v1_purges__purge_id__get",
        "parameters": [
          {
            "name": "purge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purge Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purges/{purge_id}/acknowledgements": {
      "post": {
        "tags": [
          "purges"
        ],
        "summary": "Acknowledge a purge from an edge node",
        "description": "Record that one node has acted on the purge.\n\nIdempotent by primary key: an agent retrying after a timeout re-sends the same\nacknowledgement without inflating the campaign's progress counter.",
        "operationId": "acknowledge_purge_api_v1_purges__purge_id__acknowledgements_post",
        "parameters": [
          {
            "name": "purge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purge Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeAcknowledge"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_PurgeRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PurgeRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PurgeRead]"
      },
      "PurgeAcknowledge": {
        "properties": {
          "node_id": {
            "type": "string",
            "format": "uuid",
            "title": "Node Id"
          },
          "result": {
            "$ref": "#/components/schemas/PurgeResult",
            "default": "purged"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "PurgeAcknowledge"
      },
      "PurgeCreate": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/PurgeScope"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "region_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "minLength": 2,
                "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
                "examples": [
                  "eu-central"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Code"
          },
          "path_pattern": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Pattern"
          },
          "requested_by": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Requested By",
            "examples": [
              "ops@cdn.net"
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "requested_by"
        ],
        "title": "PurgeCreate"
      },
      "PurgeRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "scope": {
            "$ref": "#/components/schemas/PurgeScope"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "path_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Pattern"
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/PurgeStatus"
          },
          "target_node_count": {
            "type": "integer",
            "title": "Target Node Count"
          },
          "acknowledged_count": {
            "type": "integer",
            "title": "Acknowledged Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scope",
          "asset_id",
          "region_id",
          "path_pattern",
          "requested_by",
          "reason",
          "status",
          "target_node_count",
          "acknowledged_count",
          "created_at",
          "completed_at"
        ],
        "title": "PurgeRead"
      },
      "PurgeResult": {
        "type": "string",
        "enum": [
          "purged",
          "not_found",
          "failed"
        ],
        "title": "PurgeResult"
      },
      "PurgeScope": {
        "type": "string",
        "enum": [
          "asset",
          "region",
          "global"
        ],
        "title": "PurgeScope"
      },
      "PurgeStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "partially_failed"
        ],
        "title": "PurgeStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Problem": {
        "title": "Problem",
        "description": "RFC 9457 problem details.",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ]
      }
    }
  }
}